Creating passwords
(1) Prompt the user to enter two words and a number, storing each into separate variables. Then, output those three values on a single line separated by a space.
Ex: If the input is:
yellow
Daisy
6
the output after the prompts is:
You entered: yellow Daisy 6
Note: User input is not part of the program output.
 
(2) Output two passwords using a combination of the user input. Format the passwords as shown below.
Ex: If the input is:
yellow
Daisy
6
the output after the prompts is:
You entered: yellow Daisy 6

First password: yellow_Daisy
Second password: 6yellow6

 
(3) Output the length of each password (the number of characters in the strings).
Ex: If the input is:
yellow
Daisy
6
the output after the prompts is:
You entered: yellow Daisy 6

First password: yellow_Daisy
Second password: 6yellow6

Number of characters in yellow_Daisy: 12
Number of characters in 6yellow6: 8