Exact change
Write a program with total change amount as an integer input, and output the change using the fewest coins, one coin type per line. The coin types are Dollars, Quarters, Dimes, Nickels, and Pennies. Use singular and plural coin names as appropriate, like 1 Penny vs. 2 Pennies.

Ex: If the input is:
0 
(or less than 0), the output is:
No change
 
Ex: If the input is:
45

the output is:
1 Quarter
2 Dimes